Added function for renaming/moving files that avoids race condition
For users who keep their files in a Dropbox folder or have other continuously-synching service, program could throw "cannot access file" errors if the file was being synched just as it was being renamed/moved.

Replaced the raw call to os.rename with a function that waits 1, then 5, then 10, then 30, then 120 seconds (as all the files involved are very small, any time such a conflict arises, it should resolve in very short order). On the final wait, the user will be notified via console (in case they have the file open in a hex editor or something).